+++ packages/banana-react/src/message/index.ts
@@ -1,0 +1,3 @@
+import { BMessage } from 'banana-ui';
+
+export const Message = BMessage;