+++ packages/banana/src/index.ts
@@ -8,6 +8,7 @@ import BInput from './input';
 import BMarquee from './marquee';
 import BMenu from './menu';
 import BMenuItem from './menu-item';
+import BMessage from './message';
 import BOverlay from './overlay';
 import BPopup from './popup';
 import BProgress from './progress';
@@ -24,6 +25,7 @@ export {
   BMarquee,
   BMenu,
   BMenuItem,
+  BMessage,
   BOverlay,
   BPopup,
   BProgress,