+++ packages/banana-react/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @banana/banana-react
 
+## 1.12.0
+
+### Minor Changes
+
+- New component: message
+
+### Patch Changes
+
+- Updated dependencies
+  - banana-ui@1.12.0
+
 ## 1.11.0
 
 ### Minor Changes